Use gson for incoming messages
This commit removes all vestiges of the old json library and it now only
uses the gson library. This decouples the code from json and makes it
easier to move code around. It is now possible to start cleaning the
rest of the code.

There is however a small problem and that is we might get into some null
pointers along the way here... need to look into that before this goes
into production.